Covid Project Tableau Visualization
The link to my Covid Tableau Visualization is attached.  It includes a map showing all the countries and the percent infected by country.  Drag your cursor over each country to see what percent of the population was infected. It also shows a bar chart of the total deaths by continent, a table of total deaths and death percentage in the world, and a line chart of selected countries and the percent of their population that was infected. Note: This is my first data project. I created this project by following a YouTube video by Alex the Analyst. This project was created as a chance for me to learn and implement Tableau when given a specific data set.